비밀번호 판단후 수정 / 삭제 기능 완료 Controller / Service 수정 완료

+++ src/main/java/com/example/project1/service/MemoService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.PathVariable;
 
+import java.sql.SQLOutput;
 import java.util.List;
 import java.util.Optional;
 
@@ -42,17 +43,26 @@ public class MemoService {
         if (requestDto.getPasswords().equals(memo.getPasswords())) {
             memo.update(requestDto);
             return memo.getId();
-        }
-        else{
-            System.out.println("비밀번호가 일치하지 않습니다");
+        } else {
+            System.out.println("비밀번호가 일치하지 않아 수정할 수 없습니다");
+            return memo.getId();
         }
 
-        return id;
+
     }
-        @Transactional
-        public Long deleteMemo (Long id){
+
+    @Transactional
+    public Long deleteMemo(Long id, MemoRequestDto requestDto) {
+        Memo memo = memoRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("아이디가 존재하지 않습니다")
+        );
+        if (requestDto.getPasswords().equals(memo.getPasswords())) {
             memoRepository.deleteById(id);
-            return id;
+            return memo.getId();
+        } else {
+            System.out.println("비밀번호가 일치하지 않아 삭제할 수 없습니다");
+            return memo.getId();
         }
     }
+}
